Add the files generated by javasphinx in the git
That way, we don't need that deprecated piece of code. Plus, the java
code is near to dead so it should not evolve in the future.

If I am to add new java bindings (for s4u), I'll write the doc weaved
into the current doc that already contains the C++/Python/C doc. So I
won't use javasphinx for that either.

As a bonus, I may be able to generate the doc in the Debian package.