smpi: many classes died tonight, but that will save kitten on the long term.
smpi::Coll was a class with 11 static methods (one per MPI function
that has several implementations: allgather, allgatherv, etc) and
nothing else.

That class was derivated for each implementation of a given MPI
function, and the resulting class was overloading only one of the
methods while the others were ignored. There was well over 100 such
child classes.

The overloaded methods were static in their class, and always used as
a function (the class was never instanciated).

All this was written with some macros to reduce the burden.

Instead, we now use regular functions in the smpi namespace.