Inventive hack to get flex raising exceptions on parse error
The core of this (over 2500 lines) patch reads as follows:
-    fprintf(stderr, "%s\n", flexml_err_msg);
+    surf_parse_error(flexml_err_msg);

But the code we want to change is generated by Flex, so ... I have to
augment the perl script that rewrite some parts of the Flex result for
that purpose. Because yes, we do have such a perl script.

At the end, we need to regenerate the flex parser with the latest
version of Flex to apply this patch, resulting in that jumbo patch.