document Virtual machines
Even if I hate to document code that should change. Here, we should
kill the fact that s4u::VirtualMachine extends s4u::Host, as it only
complicates things for nothing. It probably requires that we
dupplicate the methods this_actor::exec() and Actor::set_host() to add
a variant taking a VM instead of the host, but that would still be an
improvement.

Also, the zone C type should probably be renamed into netzone for
consistency.