Clearly state that we don't care about the return code of actors
Earlier, the actor had the exact same prototype than the C main()
function. In particular, it was supposed to return an integer. But
that value is ignored in simgrid since years (or maybe since ever).

So, change the expected return code to 'void' to make this clear and
reduce the clutter in user code.

The deprecation code should be in place until v3.30 for s4u code.

The MSG clients are not modified, and the MSG implementation is only
slightly modified with a crude cast (that mandates the
-Wno-error=cast-function-type flag on the command line when compiling
MSG).