Make sure that the doc of the sg_actor functions land in the doc
At some point in the past, I came to the conclusion that I did not
want doxygen to introspect the implementation of the code but only the
header files. I think it was because I did not want doxygen to grab
and publish the internal symbols used for the implementation, but only
to document the public interface. I'm not entirely sure because it was
a long time ago :-/ Maybe I'm missing another issue that will resurface.

But this policy requires that the docstrings are in the header, while
good practices advise to keep docstrings close to the implem, to
increase the chance of getting them updated when the implem changes.

Now that we use autodoxy, we manually pick which symbols get
displayed, so it does not matter anymore to hide stuff from Doxygen as
when we had automatic content listing.

I don't feel (yet) like moving the docstrings from the headers back to
the implem a few years after moving them in the other direction, but
since the docstrings of the C implem is close to the implem (as it
should), let's show that implem to doxygen so that the C symbols get
documented in the refguide.

And yes. This documents a one line commit with 20+ lines of logs. This
will allow agier to point me to that commit when I change my mind in
the future ;) eg when the issue that I forgot will resurface.