ns3: get ride of the node_num field, that was only used to obfuscate the debug messages
This forced my to change the signature of ns3_add_direct_route() to
take routing::NetPoint instead of NetPointNS3, to use get_cname() on it.

Also, in a cluster, I had to connect the private link of each host to
the actual router of the cluster instead of to a fantom netpoint. But
this seems to be a greater good, actually.