Revert "Assert that the log categories are unique"
The proposed method is not reliable, because when log categories are
not unique, there is 2 variables of the same name and 2 constructors
of the same name.

Redefinitions should thus be detected by the linker and we cannot fix
it reliably from our side.

I'll use less common category names in the library to reduce the risk
of conflict.

This reverts commit 7ca53b5ebe9b96f24832b6890a34e13599b254c2.