Sanitize how tracefiles are handled
In particular, SMPI was always providing a trace file per actor, even
when only one file was provided, with the content for all actors. That
was a waste of memory and processing.

But the worst was that the same file was opened N times, limiting this
solution to 1024 ranks on linux and 256 on Mac OSX. We could have
asked the users to increase the limit of file descriptors per process
to overcome this limitation.

Hopefuly with this change, the same file will be opened once only,
removing this limit all together.

Fixes https://framagit.org/simgrid/simgrid/-/issues/62

This cleanup also impacts S4U code, that cannot access the
simgrid::xbt::action_fs global anymore to open the shared file itself,
but must use xbt_replay_set_tracefile() for that.