New example: Torus cluster of multi-cpu hosts(Star Zones)
Add a new example showing how to build a torus cluster composed of
netzones instead of single hosts.

Each leaf is a StarZone, representing a host with 8 CPUs connected in a
star. CPU-0 is the gateway for each StarZone.

Runs a simple broadcast application where each CPU runs a specific
actor.