Fix FatTreeZone with limiter links
Are we sure we want this "feature" in complex clusters?
As for Torus, no test using it.

Segfault when communicating with limiter links.
Routers inside the FatTreeZone don't have limiters, even when user sets
the limiter on the XML.

Code was adding nullptr to the list of links in this case, generating a
segfault in communicate function.

Example:
  <zone id="world" routing="Full">
    <cluster id="bob_cluster"
	     prefix="node-" radical="0-15" suffix=".simgrid.org"
	     speed="1Gf" bw="125MBps" lat="50us"
             topology="FAT_TREE" topo_parameters="2;4,4;1,2;1,2"
	     loopback_bw="100MBps" loopback_lat="0" limiter_link="100MBps" />