Mix create_host and add_route during platf creation.
Allow creating hosts and adding CPUs during platform description.

With the C++ interface, it's easy to mix the creation of hosts and
routes in the same for loop. However, the old code had the assumption
that all hosts were created when we added new routes. This info was used
to initialize the routing_table.

Change code to resize the routing table when necessary (number of
elements in the zone has changed).

Zones with the error: Full and Floyd

Added UTs for this case.