ClusterZone placeholder and kill some old code
Put the old ClusterZone class as a placeholder for other "Cluster"
classes.

Keep the same behavior of Engine::get_filtered_netzones for wrench
projects.

The old ClusterZone is now BaseCluster and it's used to implement the
"complex" Clusters (Torus, DragonFly and Fat-Tree).

Kill old code that doesn't make sense for these netzones