New implementation for bandwidth factors
Bandwidth factors are now implemented at the Action level, reducing the
speed that an action advances (e.g. the number of bytes transmitted).

In the past, the factor took place at the maxmin system, limiting the
amount of resources a communication could use.

For example, a bw factor of 0.97 (default for LV08 network model) was
reflected by reducing the link capacity to 0.97*C. So, a 100MBs link had
97MBs capacity in the maxmin system.

Now, a communication alone using this link may use the 100MBs, but after
one second, it'll transmit only 97MB of data.

NOTE: This change may impact the timing of your experiments.