Adjust timing of SMPI tests
For sure SMPI is the most impacted by the bandwidth factors changes.
It seems especially impacted when collective comms are involved.

The old version used 2 factors for SMPI comms:
1) network/bandwidth-factor was used to reduce the link capacity (e.g.
0.97*C for LV08)
2) smpi/bw-factor was used for each communication, limitating the flow
capacity.

Now, the code is simplified, each communication has only 1 bw-factor
that is applied after it's done, at the update remaining phase.
In most cases, only the bw-factor are applied now but after the comm is
done, not before