Modeling hints: I/O operations
Adds an example showing how to model I/O operation in SimGrid.
It uses data from real experiments on Grid'5000 to model the impact of
concurrent operation on disks and the variability/noise in single I/O
operations.

Adds an entry in the doc presenting the results.
The code is available at: docs/source/tuto_disk.

The rst file was generated from the analysis.org file.
The folder includes a dockerfile to rerun the experiment if needed.