I/O factors: noise for disks.
Implements variability in disk operations.
Similarly to network factors, the effective amount of bytes read/written
in disk operations is scaled by a factor defined by the user.

Users can define a different callback for each disk. The callback is
called when SimGrid updates the remaining amount of bytes in disk
operations.