Reorganize VM code
After 8 years of probation, let's consider VM as an actual part of S4U
rather than as a plugin. This change should be transparent for the
users.

+ Only the Live Migration and its dirty page tracking mechanisms remain
plugins.
+ The user interface has been moved to src/s4u and the kernel
counterpart to src/kernel/resource.
+ The simgrid::vm namespace doesn't exist anymore, but the live
migration stuff is in simgrid::plugin::vm
+ The contents of VmHostExt.* has been merged to
s4u_VirtualMachine.cpp and VirtualMachine.hpp as this extension is
always needed.
+ Cleanup a bit the namespaces too