Have the application execute its transition before returning SIMCALL_EXECUTE_ANSWER
This makes the RPC protocol somewhat more complex, but it's mandatory
to get the updated observer.

This is useful in particular to Isend/Irecv observers that have a
set_comm() field that cannot be set before executing the simcall.
CommDet need that field anyway, so let's change the protocol that way.