Ensure that the monkey tesh passes while not hindering the debugging pleasure
It fails because it displays the full paths, that are only valid on my
machine. I could cut the paths from the output, but then it'd be
somewhat more complex to rerun the test when it fails (now, you just
copy/paste the exact output).

So I ask tesh to display but ignore the output.