fix weird bug on pstate changes with VM
at VM creation we did this:

/* We create cpu_action corresponding to a VM process on the host
operating system. */
action_ = physical_host_->get_cpu()->execution_start(0,core_amount_)

but in the version of execution with 2 parameters, the second one is a
user_bound not the desired amount of cores!!
This leads to entering a special in Cas01::on_speed_change and change
the bound to 1 instead of the new speed! (cpu_cas01.cpp:92)