Reimplement s4u::Barrier natively, and make them visible from MC
It comes with a DPOR dependency computation and a s4u hijacked test.

The ordering of some simcalls changed in SMPI, as there now much less
of them when using barriers. This impacts in particular the order in
which PID are given in the gh-139 test. Adding a link here for reference:
https://github.com/simgrid/simgrid/issues/139

MSG_barrier_destroy() cannot take a const barrier anymore, but how was
it possible in the first place?