Comment back the unsuccessful try to use faster matrix solver.
Ideally we would like to use PartialPivLU most of the time. But in some
cases, as tested in the UTs, we need the complete pivoting and the
assurance given by FullPivLU.

Document why it didn't work.
Need another way to check that the solution is invalid that doesn't
include isNaN.