Make sure that actors killed by a timer are properly finished
ActorImpl::exit() does not schedule the victim for execution because
exit() usually denotes a suicide.

When ran from a timer, it's executed in maestro context, so the victim
needs to be scheduled to be properly freed.

It worked properly before commit 2cb8c87fc8 because the actors were
completely destroyed by Context::stop() which is still used here. Now,
the cleanup is split in 2 parts (in actor context and in maestro
context), that are not directly inter-linked.

Gosh, that was a long debug session for a small patch. I hope it's OK now