Kill useless complexity in mmalloc: one of the dlsym
dlsym is not necessary to retrieve the LD_PRELOAD-injected symbol.

The killed code was intended to be used in libsimgrid, trying to
detect when libsgmalloc was already in memory to use that copy
instead. But that's useless: LD_PRELOAD ensures that the copy of the
injected version is the only existing one in the process.

In addition, that may resolve stability issues that my previous commit
didn't solve as nicely on the build daemons than on my laptop with gcc.