Remove the need of pthread_mutex in mmalloc, to allow its use with sthread
Historically, mmalloc was used in all case when MC was activated at
build time (even if MC was not involved at runtime), forcing this
protection against race conditions for e.g. parallel simulations.

But nowadays, mmalloc is only activated at runtime in the verified
application. In other runtime setups, mmalloc uses the libc's malloc
implem instead of itself (that's why we use dlsym in the
initialization of mmalloc, to find the replacement).

So, removing this code should have no impact whatsoever.
Only maybe speeding up the verification a tiny bit.

And now, sthread becomes usable in conjunction with Mc Simgrid!!!

Given the current poor coverage of the pthread API, we can't verify
any useful program, but at least we have a valid (and rather robust)
prototype of such a verification tool.