Ensure that the verified application still works if get_max_consider() is not called at a given depth
This can happen after a backtrack: the Checker already knows how many
times this can be considered, but the App was reset in between so it
forgot.

So, make get_max_consider() const, and do the initialization in the
observer constructor, where it belongs. Too bad for the tiny little
tiny loss of performance compared to the lazy initialization.