setter function only need a simcall in MC or with parallel execs
This change introduce a new kind of simcalls alongside
simcall_answered and simcall_blocking: simcall_run_object_access that
is dedicated to object setter functions.

This is an answered simcall if running in // or in MC, and just a
regular function call if runnning a plain simulation.

This may speed up large simulations with a huge amount of activities,
but the perf improvement is not evaluated yet.

The performance could be further improved with a SIMGRID_HAVE_PARALLEL
option allowing the compiler to know beforhand that the simcall branch
will never be taken. But such an option would require to be tested and
I'm too lazy for that.