Completely revise the way to deal with Streamed I/Os
Do not implement a new funky HostModel but handle the io_stream call
in CLM03. Basic idea (thx alegrand): start from a regular
NetworkAction and add some extra constraints if it is an I/O stream.

+ Pros: easy to implement and inherit of all the complexity of the
network part. This would have been too much work and redundancy with a
new model. No extra flag, just a new type of I/Os (similar to the
Comm::sendto spirit)

+ Cons: had to cheat a bit to NOT inherit some of the network model
(LV08) complexity for the I/O part (bandwidth factor). Still has to
find a way to not pay the network latency if the stream starts from a
disk. Had to add a little hack to reserve more space in communicate
for the extra constraints.

This new type of I/O is tested in teshsuite/s4u/io-stream (with the
basic CM02 model without crosstraffic to verify if bottlenecks are
correctly handled, and then with the default config)