Integrate a bit of the old FAQ in the modern doc, and reword Models.rst
The integrated part was about the CM02 network model, which
complements nicely the exiting doc about the TCP models, as it only
described LV08 so far.

While at it, I reworded and reorganized this documentation chapter. In
particular, I moved made the page sectionning easier to navigate, and
I moved the links to the non-performance models (routing and MC
compute model) as an annex of this page instead of having them
breaking the flow of the chapter introduction.