Add patch for arm64 Ubuntu 22.04 in UnwindContext
The `mcontext_t` struct on Ubuntu 22.04
running on arm64 is missing the `fregs`
field. This causes a compilation failure
in `src/mc/inspect/mc_unw.cpp` since the
code there assumes the field exists on
linux-arm64.

This commit adds a new CMake variable
SIMGRID_PROCESSOR_arm64 to account for
the additional architecture more
explicitly and to better fit the context
surrounding the code where the fix was made