Use std::unordered_set instead of std::set for EventSet
The ordering of the events in the set does not
matter in an EventSet. The std::unordered_set construct
is thus a better fit (std::set orders its elements).
While this disallows the use of the nice functions
in the standard library for forming unions etc.,
it is trivial to implement the functions `make_union()`
etc.