Allow actor transitions to be shared
Transitions that were stored in `ActorState`
prior to this commit were stored as `std::unique_ptr`
instances. But UDPOR events need to hold
references to the transitions that they are
"associated with". While UDPOR will discard events
at the appropriate time, we want to make sure the
reference to the transition in some `ActorState`
(ultimately in some `State` instance) remains
valid event after UDPOR discard the State object.
This could e.g. be helpful while debugging