"Finalize" interface to UnfoldingEvent
The interface to the UnfoldingEvent was carried over
from the previous implementation of UDPOR in the
"tiny_simgrid" repository. However, the implementation
needed to be adjusted to better suit how SimGrid is
going to implement UDPOR. Instead of holding a string
tag identifying a transition, each UnfoldingEvent now
directly holds a reference to the transition from the
state prior.

This caused some changes to need to be made in the
UdporChecker.cpp. These changes have moved the pseudocode
of UDPOR one large step closer to the actual implementation
that will be brought to life over the next few weeks