Add tests for the Unfolding object
The Unfolding object currently doesn't have much functionality,
so there is not much that needs to be tested with it. It is
effectively a wrapper around a collection of std::unique_ptr.
Eventually, more functionality may be added when more of the
implementation of UDPOR is added, although its use will still
be quite limited even later