Introduce extrinsic equivalence between events
Extrinsic equivalence between two UnfoldingEvents
simply means that we consider to events to be
"the same". This equivalence is used when UDPOR
creates a new event to determine if in fact that
event has already been created. This requires determining
the equivalence between the transitions associated with
those events as well as the dependencies of each of
the events.