Add topological sort of configuration events
Topological sorting is an important first step
in determining which groups of events in a
configuration need to be tested together when
looking at all possible subsets of maximal events
of a configuration. The idea is that you first
build a compatability graph (soon to come) whose
elements represent "chains" of the event structure
that are mutually exslcusive and whose edges
describe which chains are mutually compatible
with one another. Then, you look for all possible
cliques within the graph to determine all possible
configurations of these chains. Each of the
major components of this process will be added in
subsequent commits