Add complicated topological sort test
The test added ensures that for each element
in the sequence of events in the topological
sorting:

1. no event in the history of event `i` appears
in the list of events before event `i` (i.e.
events 0, ..., `i - 1`) in the case of the reverse
graph

2. every event in the history of event `i` (except
`e` itself) appears in the list of events before
`i` (i.e. events 0, ..., `i - 1`) in the case of
the event structure ordering