Sanitize the handling of timeouts in CommImpl, ExecImpl and IoImpl
There is no need for a timeout detector in ExecImpl and IoImpl, as
wait_for injects a Synchro to that extend on need.

Use a uniq_ptr on the CommImpl timeout detectors, so that the memory
is managed automatically. As it used to be for ExecImpl.

This kills the last occurrence of surf in the code. For real this time.