Add "working" (but untested) implementation of iterative subsets
The k-subsets higher-order iterator appears to be
working as expected after fixing a couple of issues
while run under GDB. The next phase is to test this
iterator heavily on a number of different examples
to prove that the iteration works as expected.

Future commits will add two important data structures,
viz. the LazyPowerSet and the LazyKSubsets, both of
which will be used when computing cliques in the
compatibility graph later on