Remove CompatibilityGraph and friends
The CompatibilityGraph and friends that were
added in order to compute maximal sets of events
needlessly complicated the algorithm. A much
better approach was suggested which will be
implemented in the next phase.

While perhaps a rebase could also have been
performed, it would have required quite careful
precision. Since the end goal is the same and
there are better uses of our time, I simply opted
to tag on another commit