Use boost::iterator_facade for History::Iterator
The iterator that is used in the `History` class
was adjusted to use boost::iterator_facade, as it
simplifies life considerably, makes the code more
readable, and provides better conformance to being
a forward iterator (i.e. it probably handles some
subtleties with forward iterators etc. and implements
many methods automatically)