Add working implementation of maximal_subsets_iterator
There were a few subtleties in computing traversing
the graph of maximal subsets iteratively. The most
important was the distinction between the case of
expanding the current maximal event set and attempting
to find the newest branch point (i.e. walking back up
the tree after performing the depth-first order).